sys-kernel: fix kernel build path issues to build coreos-firmware
In order to build `coreos-firmware` correctly, we need to set
`$KERNEL_DIR` and `$KBUILD_OUTPUT` before doing `inherit linux-info`.
Otherwise `$KV_FULL` ends up being an invalid path, and as a result,
`coreos-firmware` becomes an empty package.

Also we should make `coreos-firmware` depend on coreos-source and
coreos-modules, to be able to get a correct list of kernel modules.