app-emulation/containerd: make containerd listen on localhost
In the past there had been an issue of containerd listening on a public
address with TCP port 10010. The issue was fixed in all channels, by
disabling the gRPC plugin of containerd. [1]  However, Flatcar edge
enabled the gRPC plugin again, to support cgroup v2 for containerd. [2]
As a result, the original issue of containerd happened again on Flatcar
edge.

To fix that, we should make containerd listen on localhost 127.0.0.1,
and a TCP port 10010. See also containerd/cri docs. [3]
Also bump containerd to 1.1.2-r3.

[1] https://github.com/coreos/bugs/issues/2524
[2] https://github.com/flatcar-linux/coreos-overlay/commit/8184af2518634c115dd6ef623959da5948be4cca
[3] https://github.com/containerd/cri/blob/53c2230ec09d3672dbfb39756954b1a4afdb1ab8/docs/config.md