coreos-devel,sys-apps: do not update revisions for mantle and ignition
The PR https://github.com/flatcar-linux/coreos-overlay/pull/65 caused
regressions that the git sync action fails. It happens because the PR
incremented the revision of each ebuild, like `-r1` or `-r2`. We should
not do that, because it will now try to apply the diff on the rename to
each build-* branch, but some build branches do not have the exact
ebuild name. In case of mantle, the diff tries to find
`mantle-0.13.0-r1.ebuild`, but build-2191.x.0 has only
`mantle-0.12.0.ebuild`.

Let's revert the rename part. Then git-sync will not try to apply the
rename patch to each build branch.