sys-kernel: Update coreos-firmware commit reference to match the Manifest
When the coreos-firmware Manifest was updated in
e0e285c0bf7a6953fafbe10a5539bc8c84dbf48f
the commit reference in coreos-firmware-99999999.ebuild
was still pointing to the old file.
This led to a failed download verification.

Use the commit ID which the Manifest expects to be used
for the download.