profiles: enable sssd only for sudo on amd64 to avoid talloc failures
Jenkins pipeline for arm64 started failing with `sys-libs/talloc`,
which depends on the build system `waf`. When running `configure`
script, `waf` tries to run a cross-compiled binary, which fails with
`Exec format error`.

Unfortunately waf is not able to deal with running tests with
cross-built binaries, when no qemu-aarch64-static is installed on the
host system. Since the Jenkins worker is a minimal Flatcar system, we
cannot simply install a qemu static package on the system. Also there
is no simple way to disable running such tests in talloc and waf.

So for now, we should enable sssd for sudo only for amd64, not for
arm64. Then sudo for arm64 will not pull the new dependencies like
talloc.