Revert "app-admin/kubelet-wrapper: replace rkt with docker"
It is not simple for kubelet-wrapper to use docker, because
upper layers like Lokomotive or Typhoon still rely on rkt-specific flags
and env variables. So let's revert it back to rkt, so we can make
kubelet-wrapper work again with the existing code.

Note, in the long run, we should figure out how we can make the wrapper
work with both docker and rkt, at least for a transition period.

This reverts commit 8a03df3ede2f696f4f6ae03931ff62851c8a37fb.