coreos-base/oem-gce: Bind mount nsswitch.conf for /etc/hosts
The host's /etc/nsswitch.conf is a symlink to
/usr/share/google-oslogin/nsswitch.conf
but that is not present in the rkt container.
Do not only bind-mount /etc but also the target
of the symlink. With a broken nsswitch.conf
any entries in /etc/hosts are not considered
which makes problems when a custom DNS server
is used.