coreos-base/oem-rackspace: Get public and private IP based on routes
The OEM package does not set net.ifnames=0 in the kernel command line
(as, for example, oem-ec2-compat does) to ensure that ifnames are ethX.

Regardless of the ifname scheme, get the public and private IPs
based on the routes that get set up or by the prefix they have.
https://support.rackspace.com/how-to/basic-network-troubleshooting/
shows that the private IPs (ServiceNet) are 10.x.x.x.