.github: detect kernel version correctly
`kernel-apply-patch.sh` cannot detect the existing kernel version,
if the version does not have a patchlevel, e.g. `5.6`. So the old
kernel version variable becomes an empty string, and the final pull
request has an empty field after the `from` string.

If the Manifest does not have a `patch-` line, try to read a `linux-`
line again, to detect the correct kernel version.