.github: run weekly once to check for usual packages
We do not need to run once in a day to check for updates from
ordinary packages. Most releases happen once in more than a week.
So schedule the Github actions only once in a week for most packages.
Go on Mon, Rust on Tue, Docker on Wed, Runc on Thu.

Note, we still need to check for Kernel once in a day, as Kernel
releases happen quite often.