coreos-base/oem-vmware: Use flatcar.autologin for the console
The user can only login via SSH if no password was set and can't login
over the VGA console in the web UI. To login the user needs to press
reboot and then append flatcar.autologin to the kernel command line parameters
in GRUB each time. The user may also not know that this option even exists.
Set flatcar.autologin by default in the kernel command line parameters for
VMware so that users don't need to set this themselves.