profiles: Disable ccache
The unzip update in the portage-stable branch going along with this PR
suddenly fails to compile because ccache permissions are wrong in one
subfolder.
Disable ccache because it only gives a low hit rate anyway and once a
package is compiled, emerge will reuse the binary package. (A possible
compilation performance regression would be if a kernel patch is tested
and the kernel package needs to be build over and over again without being
able to keep the object files - not sure if this or something similar is
often the case.)