sys-apps/bpftool: update to 5.7 to adjust it to Kernel 5.7
The bpftool has relied on an experimental branch for a long time.
Now the bug fixes are already include in the mainline Kernel 5.7, so
it is possible to simply rely on mainline.

So download a mainline Kernel tarball, unpacks only necessasry parts
from that, and build it by changing several build flags.

Also make it depend on elfutils with static-libs, as to avoid occasional
build failures due to missing `libelf.a`.