coreos-base/google-oslogin: Update to 20200910
This change updates to the latest oslogin version provided by Google.
Since our last update, this was split into a different repo and the
directory structure changed significantly.

It also added group support, which needed to be added to the
nsswitch.conf file that we ship.

Flatcar users require docker group permissions, so ensure oslogin gives
that permission by shipping a separate group.conf file that gets
installed when oslogin is enabled.