x11-drivers/nvidia-drivers: Fix the dev container download URL
The kola tests fails to download during the release because the
artifacts of the release has not been pushed to the website yet.
This adds the logic to check if the URL is 200, then only download
or else fallback to the GCS bucket url.

This commit also changes a bug with the check to see if nvidia
is installed or required.

Signed-off-by: Sayan Chowdhury <sayan@kinvolk.io>