app-emulation/amazon-ssm-agent: cross-compile and use correct go version
The ebuild was missing a call to go_export() which exports GOARCH, and so was
always built for host architecture. While COREOS_GO_VERSION was specified as
go1.12, src_compile() has to use '${EGO}' to make use of it, so we were
building with go1.16 (latest).  Upstream builds with 1.12 for this version, so
we will do the same.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>