core-base/oem-rpi4: introduce OEM config for RPI4
RaspberryPI 4 with UEFI firmware is supported by Flatcar. But the console
settings and autologin need tweaking to make it usable, so introduce a config
that we can rollout with a new image type. The OEM config also includes the
UEFI firmware bits, so that the resulting image can be installed to a USB stick
and directly booted. This requires recent recent (<1 year old) EEPROM,
BOOT_ORDER changes and some documentation and then it works.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>