bootstrap: add `minimal` useflag to BOOSTRAP_USE
with the recent update of `dev-lang/perl`, we added the `minimal`
useflag.

This one is not taken in account from `package.use` into the stage 2 of the boostraping,
because we do an `export USE=...`.

Following the precedence of the USE flag with Gentoo, the `export` will
be used in first, so the `package.use` with our `dev-lang/perl minimal`
won't be used.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>