sec-policy/selinux-base: add missing BDEPEND=python[xml]
selinux-base requires python to generate xml files, but the dependency
is implicit (through policycoreutils). Flatcar made that dependency
conditional on USE=python in policycoreutils so that we don't include
python in our images, but this causes selinux-base to fail depending on
ordering in the bootstrap process.

Fix that failure by addin an explicit dependency.