sys-kernel/coreos-modules: explicitly enable EFI_GENERIC_STUB_INITRD_CMDLINE_LOADER
We rely on this setting to make iPXE booting work on EFI platforms. In iPXE we
use 2 initramfs': the kernel builtin one and a pxe specific one that contains
the contents of the usr partition. This appears to rely on the EFI stub, which
unpacks the second one based on the passed commandline parameter (initrd=).

This affects arm64 kernels after v5.11 because of this commit:
https://github.com/torvalds/linux/commit/6edcf9dc2e1aff3aa1f5a69ee420fb30dd0e968a

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>