.github: find latest calico tag from tigera operator
The mirror-calico workflow has been failing because it currently determines
version=v3.22.0-0.dev-typha, which is not the tag used by the individual
container images. Rewrite the version logic to determine the version based on
what is in the tigera operator manifest. This is the same manifest that we use
to deploy calico in mantle.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>